+++ src/osd/CMakeLists.txt
@@ -11,9 +11,6 @@ endif()
 
 set(osd_srcs
   OSD.cc
-  pg_scrubber.cc
-  scrub_machine.cc
-  PrimaryLogScrub.cc
   Watch.cc
   ClassHandler.cc
   PG.cc
@@ -24,10 +21,13 @@ set(osd_srcs
   ECTransaction.cc
   PGBackend.cc
   OSDCap.cc
+  scrubber/pg_scrubber.cc
+  scrubber/PrimaryLogScrub.cc
+  scrubber/scrub_machine.cc
+  scrubber/ScrubStore.cc
   Watch.cc
   Session.cc
   SnapMapper.cc
-  ScrubStore.cc
   osd_types.cc
   ECUtil.cc
   ExtentCache.cc
