+++ src/osd/scrubber/PrimaryLogScrub.cc
@@ -6,8 +6,8 @@
 #include "common/scrub_types.h"
 #include "osd/osd_types_fmt.h"
 
-#include "PeeringState.h"
-#include "PrimaryLogPG.h"
+#include "osd/PeeringState.h"
+#include "osd/PrimaryLogPG.h"
 #include "scrub_machine.h"
 
 #define dout_context (m_pg->get_cct())
